requirements.txt
# Core Framework
Flask==2.3.2
Werkzeug==2.3.6

# Graph of Thoughts Dependencies
backoff>=2.2.1,<3.0.0
openai>=1.0.0,<2.0.0
matplotlib>=3.7.1,<4.0.0
numpy>=1.24.3,<2.0.0
pandas>=2.0.3,<3.0.0
sympy>=1.12,<2.0
torch>=2.0.1,<3.0.0
transformers>=4.31.0,<5.0.0
accelerate>=0.21.0,<1.0.0
bitsandbytes>=0.41.0,<1.0.0
scipy>=1.10.1,<2.0.0

# Cerebras API
cerebras-cloud-sdk>=1.0.0

# Web Development
Jinja2>=3.1.2
MarkupSafe>=2.1.3
itsdangerous>=2.1.2
click>=8.1.3
blinker>=1.6.2

# Data Processing
beautifulsoup4>=4.12.2
lxml>=4.9.3
requests>=2.31.0
python-dateutil>=2.8.2

# Utilities
python-dotenv>=1.0.0
uuid>=1.30
pathlib2>=2.3.7

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Production
gunicorn>=21.2.0
redis>=4.6.0

# Optional Dependencies for Enhanced Features
# Uncomment as needed:
# plotly>=5.15.0          # Advanced visualization
# dash>=2.11.1            # Interactive dashboards  
# celery>=5.3.1           # Background task processing
# SQLAlchemy>=2.0.19      # Database ORM
# psycopg2-binary>=2.9.7  # PostgreSQL support